Refactor quiz features into module
