+++ composer.json
@@ -6,6 +6,7 @@
       "NuclearEngagement\\Services\\": "includes/Services/",
       "NuclearEngagement\\Requests\\": "includes/Requests/",
       "NuclearEngagement\\Responses\\": "includes/Responses/",
+      "NuclearEngagement\\Modules\\": "nuclear-engagement/inc/Modules/",
       "NuclearEngagement\\": "nuclear-engagement/"
     }
   },
